build: Create a @jkcfg/kubernetes npm package
This is a bit of a weird package as we have 2 levels of APIs, one in TS, one in
JS in the same package. We can't use the whole API surface from TS, but we can
from JS.

In any case, it's a lot easier to build the package we want to publish in a
dist repository and publish that. In particular, this way, the import paths
become:

  import * as k8s from '@jkcfg/kubernetes/api'

instead of:

  import * as k8s from '@jkcfg/kubernetes/src/api'