Make sure go is installed for gen
Because we now import generated (or generated then compiled) files,
they must be generated before test is run. The Makefile takes care of
that (provided `make test` is used). However, to run the generation
step, the correct version of Go must be installed, so that must be
moved earlier in the CI script.