Add another tool to the overlay toolbox
This adds rewriteImageRefs, which will apply a function to rewrite all
image refs referred to in containers in a resource. This can be used
on its own, e.g., to rewrite all image refs to refer to a local
registry; and, will be wired into the overlay interpreter, to
implement the `imageTags` transformation.

To get there, we need a bit of extra scaffolding: a generator for
image containers, and a more general purpose object traverser. It's an
open question whether it would be better to make this use a zipper
(i.e., to construct a new object with the "deep" change, instead of
mutating). On the one hand, pure functional is better; on the other,
would make typing quite challenging.