Add util for writing a list of resources out
It's handy to be able to pass the result of e.g., evaluating an
overlay, to a procedure which will write it out to
files. writeResources takes the writer and will write the resources
out one to a file -- each named for the name and kind of the resource
and put in a directory named for its namespace.