Implement chart values from files and command-line
This fills in the ability to supply values to the chart-alike from a
file (loaded as an object, so either JSON or YAML), or the command
line.

It uses the (jk run -p) parameters `value` and `file`. The former
expects a comma-delimited list of "path.to.key=value" (but the jk
runtime won't parse these, at present), and the latter a filename.

This commit also changes the still-pretty-minimal `chart` function so
that it outs a YAML stream, instead of writing files, just so it
doesn't leave litter around. This suggested a new convenience
`writeStream`.