Provide template functions in chart module
In the spirit of making the "main" modules more convenient,
chart/index.js now exports a couple of partly-specialised functions
for loading templates from the files in a directory.

For most uses, there's only two varieties that matter: loading
templates from an input directory, and loading templates from a module
(i.e, as a module resource). In the latter case, the `resource` import
from the calling module must be supplied.