Parse the results of templates
To put the use of templates on a level with all the other ways of
constructing resources, this commit makes the chart/template module
parse the output of templates, and so return objects rather than
chunks of string.

Since any given template can return more than one resource, the
results from a directory of templates are concatenated.