Rearrange overlay and rewrite examples
This separates the (very barely) more general `overlay` from
kustomizations; these differ in that the former can be given an object
to interpret, while the latter just has a path at which to look for a
kustomization file.

In the future, `overlay` could be parameterised by the compilation
step (turning a path or object into an overlay object), but for now
it's particular to kustomizations.

This commit also rewrites the examples so they can be run with `jk
generate`, and gives `generate#valuesForGenerate` an option for
prefixing the file path. (It's easy to do otherwise, but convenient to
do here).