Use only the first segment of the schema group
The tool that generates the schemas in use here,
https://github.com/instrumenta/openapi2jsonschema, shortens group
names for Kubernetes definitions to just the first segment -- e.g.,
`apiextensions.k8s.io` gets the group `apiextensions`. The group is
then used to name the schema file generated.

This commit reproduces that logic in schema.js, so that it can find
the right file for resources that have a dotted name as above.

To be able to test the path construction in isolation, I moved things
around a little, so access to resources is done from validate.js.